+++ src/containers/shared/components/DualAxisAreaChart/DualAxisAreaChart.tsx
@@ -10,6 +10,7 @@ import {
   CartesianGrid,
 } from 'recharts'
 
+/** Configuration for a single Y-axis and its associated area series. */
 export interface AxisConfig {
   dataKey: string
   label: string
@@ -18,6 +19,10 @@ export interface AxisConfig {
   show: boolean
 }
 
+/**
+ * A recharts AreaChart with two Y-axes (left + right), gradient fills,
+ * and configurable axis formatting.
+ */
 export interface DualAxisAreaChartProps {
   data: any[]
   xAxisKey: string
