+++ src/containers/shared/components/TVLVolumeChart/styles.scss
@@ -1,4 +1,4 @@
-@use '../../shared/css/variables' as *;
+@use '../../css/variables' as *;
 
 $tvl-color: #32e685;
 $volume-color: #7919ff;
